+++ src/common/classes/display-set.class.js
@@ -49,6 +49,30 @@ export class DisplaySet {
     /**
      * @locus Anywhere
      * @access public
+     * @summary Find allowed display units
+     * @param {String} menu the target menu
+     * @param {String} user
+     * @returns {Array<DisplayUnit>} the list of DisplayUnit's allowed for this user
+     */
+    async allowedInMenu( menu, user ){
+        _trace( 'DisplaySet::allowedList() menu='+menu+' user='+user );
+        let result = [];
+        await this.enumerate( async ( name, unit, args ) => {
+            const menus = unit.get( 'inMenus' );
+            if( menus.includes( args.menu )){
+                const allowed = await unit.accessAllowed( args.user );
+                if( allowed ){
+                    args.result.push({ name, unit });
+                }
+            }
+            return true;
+        }, { menu: menu, user: user, result: result });
+        return result;
+    }
+
+    /**
+     * @locus Anywhere
+     * @access public
      * @summary Find a unit definition by name
      * @param {String} name
      * @returns {DisplayUnit} the found definition, or null
@@ -66,12 +90,16 @@ export class DisplaySet {
      *  the `cb()` function must return true to continue the enumeration, false to stop it
      * @param {Any} arg an optional argument to be provided to the cb() callback
      */
-    enumerate( cb, arg=null ){
+    async enumerate( cb, arg=null ){
         _trace( 'DisplaySet::enumerate()' );
         const self = this;
         assert( cb && _.isFunction( cb ), 'expected a function, found '+cb );
-        Object.keys( self.#set ).sort().every(( key ) => {
-            return cb( key, self.#set[key], arg );
-        });
+        const keys = Object.keys( self.#set ).sort();
+        for ( const key of keys ){
+            const goon = await cb( key, self.#set[key], arg );
+            if( !goon ){
+                break;
+            }
+        }
     }
 }