+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - No more provide a default class to the display units
+    - Remove useless RunContext.dataContext() method
 
 ### 1.2.0